Properly free array list arguments
git-svn-id: svn+ssh://svn.code.sf.net/p/cmusphinx/code/trunk/sphinxbase@11445 94700074-3cef-4d97-a70e-9c8c206c02f5

+++ include/sphinxbase/cmd_ln.h
@@ -372,7 +372,7 @@ void cmd_ln_set_float_r(cmd_ln_t *cmdln, char const *name, double fv);
 /*
  * Compatibility macros
  */
-#define cmd_ln_int32_r(c,n)	cmd_ln_int_r(c,n)
+#define cmd_ln_int32_r(c,n)	(int32)cmd_ln_int_r(c,n)
 #define cmd_ln_float32_r(c,n)	(float32)cmd_ln_float_r(c,n)
 #define cmd_ln_float64_r(c,n)	(float64)cmd_ln_float_r(c,n)
 #define cmd_ln_set_int32_r(c,n,i)   cmd_ln_set_int_r(c,n,i)
